rename class Object to Object_ for PHP 7.2 compatibility
